+++ mldnshiro-ssm-web/src/main/java/cn/mldn/mldnshiro/ssm/web/action/front/PageActionFront.java
@@ -1,0 +1,26 @@
+package cn.mldn.mldnshiro.ssm.web.action.front;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Controller
+public class PageActionFront {
+	@RequestMapping("/loginForm")
+	public String loginForm() {
+		return "login" ;
+	}
+	@RequestMapping("/pages/back/welcome")
+	public String welcome() {
+		return "back/welcome" ;
+	}
+	@RequestMapping("/logout")
+	public String logout() {
+		return "login" ;
+	} 
+	@ResponseBody
+	@RequestMapping("/unauth")
+	public Object unauth() {
+		return "【ERROR - AUTH】没有此类处理权限。" ;
+	}
+}
