+++ mldnshiro-ssm-web/src/main/resources/spring/spring-shiro.xml
@@ -14,17 +14,33 @@
 	<bean id="securityManager" class="org.apache.shiro.web.mgt.DefaultWebSecurityManager">
 		<property name="realm" ref="memberRealm"/>	<!-- 配置要使用的Realm对象 -->
 	</bean>
+	<!-- 通过配置FormAuthenticationFilter来实现内置登录的操作控制 -->
+	<bean id="formAuthenticationFilter" class="org.apache.shiro.web.filter.authc.FormAuthenticationFilter">
+		<property name="usernameParam" value="mid"/>	<!-- 定义表单参数名称 -->
+		<property name="passwordParam" value="password"/>	<!-- 定义表单参数名称 -->
+		<!-- 如果要进行登录的提交一定要有一个控制器（Servlet、Filter），但是至少需要有一个提交地址 -->
+		<property name="loginUrl" value="/loginForm.action"/>
+	</bean>
 	<!-- 配置文件里面已经将所有的配置交由Spring负责了，所以对于过滤的处理操作一定要交给指定的类完成配置 -->
 	<bean id="shiroFilter" class="org.apache.shiro.spring.web.ShiroFilterFactoryBean">
 		<!-- 在此过滤管理之中定义要使用的安全管理器 -->
 		<property name="securityManager" ref="securityManager"/>
 		<!-- 当认证或授权出现了错误之后应该跳转到登录页面，要由Action跳转到JSP页面 -->
-		<property name="loginUrl" value="/login_pre.action"/>
+		<property name="loginUrl" value="/loginForm.action"/>
 		<!-- 当授权失败之后跳转的页面路径，也是通过Action跳转 -->
 		<property name="unauthorizedUrl" value="/unauth.action"/>
+		<!-- 当登录成功之后跳转到的指定页面 -->
+		<property name="successUrl" value="/pages/back/welcome.action"/>
+		<property name="filters">	<!-- 由于现在重新定义了过滤，所以要重新追加 -->
+			<map>
+				<!-- 当执行了authc的检测的时候，明确的找到新的过滤配置类 -->
+				<entry key="authc" value-ref="formAuthenticationFilter"/>
+			</map>
+		</property>
 		<!-- 定义所有访问路径处理规则 -->
 		<property name="filterChainDefinitions">
 			<value>
+				/loginForm.action=authc
 				/pages/back/**=authc
 			</value>
 		</property>
